LayoutTests/imported/w3c/web-platform-tests/css/css-anchor-position/at-position-try-invalidation-shadow-dom-expected.txt


PASS #host is initially left: 999999px
PASS #slotted is initially left: 999999px
FAIL #host with inserted @position-try applied assert_equals: expected 100 but got 999999
FAIL #slotted with inserted @position-try applied assert_equals: expected 100 but got 999999
